Isolate logoutUser controller tests
